+++ lightkurve/correctors/tests/test_sffcorrector.py
@@ -3,7 +3,8 @@ import pytest
 import numpy as np
 from astropy.utils.data import get_pkg_data_filename
 from numpy.testing import assert_array_equal
-from ... import LightCurve, KeplerLightCurveFile, KeplerLightCurve
+from ... import LightCurve, KeplerLightCurveFile, KeplerLightCurve, \
+                TessLightCurve, LightkurveWarning
 from .. import SFFCorrector
 
 
@@ -166,3 +167,10 @@ def test_sff_breakindex():
                  centroid_col=np.random.randn(20),
                  centroid_row=np.random.randn(20), windows=1)
     assert_array_equal(corr.window_points, np.asarray([5, 10]))
+
+
+def test_sff_tess_warning():
+    """SFF is not designed for TESS, so we raise a warning."""
+    lc = TessLightCurve(flux=[1, 2, 3])
+    with pytest.warns(LightkurveWarning, match='not designed for'):
+        corr = SFFCorrector(lc)